deployment/instrument/pyproject.toml
[tool.poetry]
name = "instrument"
version = "0.1.0"
description = ""
authors = ["thy-hoang-nn <hoangthynguyennguyen@gmail.com>"]

packages = [{include = "instrument"}]

[tool.poetry.dependencies]
python = "^3.9.0"
fastapi = "^0.101.0"
uvicorn = "^0.23.2"
pydantic = "1.10.10"
python-memcached = "^1.59"
python-multipart = "^0.0.6"
loguru = "0.7.0"
opentelemetry-api = "^1.22.0"
opentelemetry-exporter-jaeger = "^1.21.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.22.0"
opentelemetry-exporter-prometheus = "^0.43b0"
opentelemetry-instrumentation-asgi = "^0.43b0"
opentelemetry-instrumentation-fastapi = "^0.43b0"
opentelemetry-instrumentation-logging = "^0.43b0"
opentelemetry-instrumentation-requests = "^0.43b0"
opentelemetry-sdk = "^1.22.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


